motopp/requirements.txt
Flask-Login
Flask-SQLAlchemy
Flask-WTF
macholib
pip
pipdeptree
mysql-connector-python
pymysql
cryptography
requests
beautifulsoup4
pandas
redis